MDBF-713 - Move scheduler helper functions
